+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * 1.2.7 (2016-07-12)
+    * HOTFIX      #2605 [CategoryBundle]      Fixed order in combination with depth
     * HOTFIX      #2600 [CategoryBundle]      fixed order of categories in content-type
     * HOTFIX      #2585 [CoreBundle]          add fixtures without purging database
     * HOTFIX      #2550 [MediaBundle]         made documents list show description on add